Generate .luaurc file for aliases instead of vscode settings
The VSCode setting has been deprecated for a while now. Now in 1.44.0,
no completions are provided at all using those settings. This brings up
the issue of the '@' symbol that's forcefully put in front of all
aliases. WoS will have to do something about it. Most likely not any
time soon, as Hexcede is gone currently.